Trait cap_time_ext::MonotonicClockExt

pub trait MonotonicClockExt {
    type Instant;
    fn now_with(&self, precision: Duration) -> Self::Instant;

    fn resolution(&self) -> Duration;
}

Extension trait for cap_std::time::MonotonicClock.

Associated Types

type Instant

A monotonic clock datapoint.

Required methods

fn now_with(&self, precision: Duration) -> Self::Instant

Similar to MonotonicClock::now, but takes an additional precision parameter allowing callers to inform the implementation when they don’t need full precision. The implementation need not make any effort to provide a time with greater precision.

fn resolution(&self) -> Duration

Return the resolution of the clock.